refactor: extract web3 verification to separate module
Moves web3 identity verification logic to its own module and renames
functions for better clarity:
- get_expected_raw_message -> get_verification_message
- verify_web3_identity -> verify_identity
Improves code organization and maintainability.
